+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/ahorro/AhorroRequestPago.java
@@ -39,6 +39,6 @@ public class AhorroRequestPago implements Serializable {
 
     @JsonProperty("numero_cuota")
     @SerializedName("numero_cuota")
-    private Long numeroCuota;
+    private Integer numeroCuota;
 
 }
