+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/movimiento/MovimientoModel.java
@@ -41,7 +41,7 @@ public class MovimientoModel implements Serializable {
 
     @JsonProperty("numero_cuota")
     @SerializedName("numero_cuota")
-    private Long numeroCuota;
+    private Integer numeroCuota;
 
     @JsonProperty("signo")
     @SerializedName("signo")
