+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/prestamo/PrestamoRequestPago.java
@@ -30,6 +30,6 @@ public class PrestamoRequestPago implements Serializable {
 
     @JsonProperty("numero_cuota")
     @SerializedName("numero_cuota")
-    private Long numeroCuota;
+    private Integer numeroCuota;
 
 }
